+++ app/static/lang/lang.fr.js
@@ -169,6 +169,7 @@ export const lang = {
   staffDownText: { text: "Élément d'un personnel vers le bas" },
   increaseDurText: { text: 'Augmenter la durée' },
   decreaseDurText: { text: 'Réduire la durée' },
+  toggleDotsText: { text: 'Activar punxat' },
   cleanAccidText: { text: 'Vérifier les @accid.ges' },
   renumberMeasuresTestText: { text: 'Renommer les mesures (test)' },
   renumberMeasuresExecText: { text: 'Renommer les mesures (exécution)' },
@@ -180,6 +181,9 @@ export const lang = {
 
   // INSERT MENU ITEM  ELEMENT DE MENU INSERTION
   insertMenuTitle: { text: 'Insérer' },
+  addNoteText: { text: 'Ajouter une note' },
+  convertNoteToRestText: { text: 'Note(s) <=> silence(s)' },
+  toggleChordText: { text: 'Note(s) <=> accord' },
   addDoubleSharpText: { html: 'Double dièse &#119082;' },
   addSharpText: { html: 'Dièse &#9839;' },
   addNaturalText: { html: 'Bécarre &#9838;' },