+++ app/static/lang/lang.it.js
@@ -109,7 +109,7 @@ export const lang = {
   openUrlText: { text: 'Apri URL' },
   openExample: {
     text: 'Repertorio pubblico',
-    description: 'Apri una lista del repertorio pubblico',
+    description: 'Apri un repertorio di pubblico dominio',
   },
   importMusicXml: { text: 'Importa MusicXML' },
   importHumdrum: { text: 'Importa Humdrum' },
@@ -144,8 +144,8 @@ export const lang = {
   notationLeft: { text: 'Notazione a sinistra' },
   notationRight: { text: 'Notazione a destra' },
   showSettingsMenuText: { text: 'Pannello di impostazioni' },
-  showAnnotationMenuText: { text: 'Pannello di annotazioni' },
-  showFacsimileMenuText: { text: 'Pannello di facsimili' },
+  showAnnotationMenuText: { text: 'Finestra di annotazioni' },
+  showFacsimileMenuText: { text: 'Finestra di facsimili' },
   showPlaybackControlsText: { text: 'Controlli di riproduzione' },
   facsimileTop: { text: 'Facsimile in alto' },
   facsimileBottom: { text: 'Facsimile in basso' },
@@ -158,16 +158,17 @@ export const lang = {
   betweenPlacementText: { text: 'Posiziona in mezzo' },
   addVerticalGroupText: { text: 'Aggiungi gruppo verticale' },
   deleteText: { text: 'Elimina elemento' },
-  pitchChromUpText: { text: 'Alza cromatica' },
-  pitchChromDownText: { text: 'Abbassa cromatica' },
-  pitchUpDiatText: { text: 'Alza diatonica' },
-  pitchDownDiatText: { text: 'Abbassa diatonica' },
+  pitchChromUpText: { text: 'Alza cromaticamente' },
+  pitchChromDownText: { text: 'Abbassa cromaticamente' },
+  pitchUpDiatText: { text: 'Alza diatonicamente' },
+  pitchDownDiatText: { text: 'Abbassa diatonicamente' },
   pitchOctaveUpText: { text: "Alza di un'ottava" },
   pitchOctaveDownText: { text: "Abbassa di un'ottava" },
-  staffUpText: { text: 'Elemento su 1 rigo' },
-  staffDownText: { text: 'Elemento giù 1 rigo' },
+  staffUpText: { text: 'Alza di un rigo' },
+  staffDownText: { text: 'Abbassa di un rigo' },
   increaseDurText: { text: 'Aumenta durata' },
   decreaseDurText: { text: 'Riduci durata' },
+  toggleDotsText: { text: 'Alterna la punteggiatura' },
   cleanAccidText: { text: 'Verificare @accid.ges' },
   renumberMeasuresTestText: { text: 'Rinumera misure (test)' },
   renumberMeasuresExecText: { text: 'Rinumera misure (esegui)' },
@@ -179,6 +180,9 @@ export const lang = {
 
   // INSERT MENU ITEM / INSERISCI VOCE DI MENU
   insertMenuTitle: { text: 'Inserisci' },
+  addNoteText: { text: 'Aggiungi nota' },
+  convertNoteToRestText: { text: 'Nota(e) <=> pausa(e)' },
+  toggleChordText: { text: 'Nota(e) <=> accordo' },
   addDoubleSharpText: { html: 'Doppio diesis &#119082;' },
   addSharpText: { html: 'Diesis &#9839;' },
   addNaturalText: { html: 'Bequadro &#9838;' },
@@ -225,18 +229,18 @@ export const lang = {
 
   // HELP MENU ITEM / VOCE DI MENU AIUTO
   helpMenuTitle: { text: 'Aiuto' },
-  goToHelpPageText: { text: 'Pagina di aiuto di mei-friend' },
-  goToCheatSheet: { text: 'Scheda informativa di mei-friend' },
+  goToHelpPage: { text: 'Pagina di aiuto di mei-friend' },
+  goToCheatSheet: { text: 'Documentatione di mei-friend' },
   showChangelog: { text: 'Registro delle modifiche di mei-friend' },
   goToGuidelines: { text: 'Linee guida MEI' },
-  consultGuidelinesForElementText: { text: "Linee guida per l'elemento corrente" },
+  consultGuidelinesForElementText: { text: "Informazioni dell'elemento corrente" },
   provideFeedback: { text: 'Fornisci un feedback' },
   resetDefault: { text: 'Ripristina impostazioni predefinite' },
 
   // panel icons/ icone pannello
   showMidiPlaybackControlBarButton: { description: 'Attiva/Disattiva barra di controllo della riproduzione MIDI' },
-  showFacsimileButton: { description: 'Attiva/Disattiva pannello Facsimile' },
-  showAnnotationsButton: { description: 'Attiva/Disattiva pannello Annotazioni' },
+  showFacsimileButton: { description: 'Attiva/Disattiva la finestra di Facsimile' },
+  showAnnotationsButton: { description: 'Attiva/Disattiva la finestra Annotazioni' },
   showSettingsButton: { description: 'Mostra il pannello delle impostazioni' },
 
   // Footer texts/ Testi di piè di pagina
@@ -249,12 +253,12 @@ export const lang = {
       '<a href="https://iwk.mdw.ac.at/impressum">Imprint</a>.',
   },
   loadingVerovio: { text: 'Caricamento Verovio' },
-  verovioLoaded: { text: 'caricato' },
-  convertedToPdf: { text: 'convertito in PDF' },
+  verovioLoaded: { text: 'Caricato' },
+  convertedToPdf: { text: 'Convertito in PDF' },
   statusBarCompute: { text: 'Calcola' },
-  middleFooterPage: { text: 'pagina' },
-  middleFooterOf: { text: 'di' },
-  middleFooterLoaded: { text: 'caricato' },
+  middleFooterPage: { text: 'Pagina' },
+  middleFooterOf: { text: 'Di' },
+  middleFooterLoaded: { text: 'Caricato' },
 
   // control menu/ menu di controllo
   verovioIcon: {
@@ -264,10 +268,10 @@ export const lang = {
   },
   decreaseScaleButton: { description: 'Riduci notazione' },
   verovioZoom: { description: 'Scala dimensioni dello spartito' },
-  increaseScaleButton: { description: 'Aumenta notazione' },
+  increaseScaleButton: { description: 'Ingrandisci notazione' },
   pagination1: { html: 'Pagina ' },
   pagination3: { html: ' di' },
-  sectionSelect: { description: 'Naviga la struttura di sezione/finale del codice' },
+  sectionSelect: { description: 'Naviga la sezione/finale del codice' },
   firstPageButton: { description: 'Vai alla prima pagina' },
   previousPageButton: { description: 'Vai alla pagina precedente' },
   paginationLabel: {
@@ -275,8 +279,8 @@ export const lang = {
   },
   nextPageButton: { description: 'Vai alla pagina successiva' },
   lastPageButton: { description: "Vai all'ultima pagina" },
-  flipCheckbox: { description: 'Passa automaticamente alla posizione del cursore di codice' },
-  flipButton: { description: 'Passa manualmente alla posizione del cursore di codice' },
+  flipCheckbox: { description: 'Passa automaticamente alla posizione del cursore nel codice' },
+  flipButton: { description: 'Passa manualmente alla posizione del cursore nel codice' },
   breaksSelect: { description: 'Definisci il comportamento degli intervalli/pagine di sistema nello spartito' },
   breaksSelectNone: { text: 'Nessuno' },
   breaksSelectAuto: { text: 'Automatico' },
@@ -303,7 +307,7 @@ export const lang = {
 
   // PDF/print preview panel / Pannello di anteprima PDF/stampa
   pdfSaveButton: { text: 'Salva PDF', description: 'Salva come PDF' },
-  pdfCloseButton: { description: 'Chiudi vista di stampa' },
+  pdfCloseButton: { description: 'Chiudi anteprima di stampa' },
   pagesLegendLabel: { text: 'Intervallo di pagine', singlePage: 'pagina', multiplePages: 'Pagine' },
   selectAllPagesLabel: { text: 'Tutte' },
   selectCurrentPageLabel: { text: 'Pagina corrente' },
@@ -312,23 +316,23 @@ export const lang = {
   selectPageRangeLabel: { text: 'Intervallo di pagine:' },
   pdfPreviewSpeedModeWarning: {
     text:
-      'Solo la pagina corrente viene renderizzata in PDF, poiché è attivata la modalità di velocità. ' +
-      'Deseleziona la modalità di velocità per selezionare tutte le pagine.',
+      'In modalità veloce solamente la pagina corrente viene renderizzata in PDF.' +
+      'Deseleziona la modalità veloce per selezionare tutte le pagine.',
   },
   pdfPreviewNormalModeTitle: { text: "Seleziona l'intervallo di pagine da salvare in PDF." },
 
-  // facsimile panel/ pannello facsimile
-  facsimileIcon: { description: 'Pannello facsimile' },
-  facsimileDecreaseZoomButton: { description: "Riduci l'immagine dello spartito" },
-  facsimileZoom: { description: "Regola la dimensione dell'immagine dello spartito" },
-  facsimileIncreaseZoomButton: { description: "Ingrandisci l'immagine dello spartito" },
+  // facsimile panel/ finestra facsimile
+  facsimileIcon: { description: 'Finestra facsimile' },
+  facsimileDecreaseZoomButton: { description: 'Riduci la finestra dello spartito' },
+  facsimileZoom: { description: 'Regola la dimensione la finestra dello spartito' },
+  facsimileIncreaseZoomButton: { description: 'Ingrandisci la finestra dello spartito' },
   facsimileFullPageLabel: {
     text: 'Pagina intera',
-    description: "Mostra l'intera pagina dell'immagine del facsimile",
+    description: 'Mostra la finestra del facsimile a pagina intera',
   },
-  facsimileFullPageCheckbox: { description: "Mostra l'intera pagina dell'immagine del facsimile" },
+  facsimileFullPageCheckbox: { description: 'Mostra la finestra del facsimile a pagina intera' },
   facsimileShowZonesLabel: {
-    text: 'Mostra riquadri delle zone',
+    text: 'Mostra i riquadri delle zone',
     description: 'Mostra i riquadri delle zone del facsimile',
   },
   facsimileShowZonesCheckbox: { description: 'Mostra i riquadri delle zone del facsimile' },
@@ -337,12 +341,12 @@ export const lang = {
     text: 'Modifica le zone',
     description: 'Modifica le zone del facsimile',
   },
-  facsimileCloseButton: { description: 'Chiudi il pannello facsimile' },
+  facsimileCloseButton: { description: 'Chiudi la finestra di facsimile' },
   facsimileDefaultWarning: { text: 'Nessun contenuto facsimile da visualizzare.' },
   facsimileNoSurfaceWarning: {
-    text: 'Nessun elemento di superficie trovato per questa pagina.\n(Potrebbe mancare un elemento pb iniziale.)',
+    text: 'Nessun elemento superficiale trovato in questa pagina.\n(Potrebbe mancare un elemento pb iniziale.)',
   },
-  facsimileNoZonesFullPageWarning: { text: 'Facsimile senza zone visibile solo in modalità pagina intera.' },
+  facsimileNoZonesFullPageWarning: { text: 'Facsimile senza zone visibili a pagina intera.' },
   facsimileImgeNotLoadedWarning: { text: "Impossibile caricare l'immagine" },
 
   // drag'n'drop / trascina e rilascia
@@ -361,7 +365,7 @@ export const lang = {
   openUrlCancel: { text: 'Annulla' },
   proposePublicRepertoire: {
     html:
-      'Accettiamo proposte per ' +
+      'Accettiamo proposte su ' +
       '<a href="https://github.com/mei-friend/mei-friend/issues/new?template=public_repertoire.md" target="_blank" >' +
       'aggiunte al repertorio pubblico' +
       '</a>.',
@@ -376,7 +380,7 @@ export const lang = {
     text: 'Clicca su "Esegui workflow" per richiedere all\'API di GitHub di eseguire il workflow sopra riportato per te, utilizzando la configurazione di input specificata di seguito. La tua codifica verrà ricaricata nella sua ultima versione una volta che l\'esecuzione del workflow sarà completa.',
   },
   githubActionStatusMsgPrompt: { text: 'Impossibile eseguire il workflow: GitHub dice' },
-  githubActionStatusMsgWaiting: { text: 'Sii paziente mentre GitHub sta elaborando il tuo workflow...' },
+  githubActionStatusMsgWaiting: { text: 'Attendi, GitHub sta elaborando il tuo workflow...' },
   githubActionStatusMsgFailure: { text: 'Impossibile eseguire il workflow: GitHub dice' },
   githubActionStatusMsgSuccess: { text: 'Esecuzione del workflow completata: GitHub dice' },
   githubActionsRunButton: { text: 'Esegui workflow' },
@@ -395,7 +399,7 @@ export const lang = {
   forkRepositoryInstructions: {
     text:
       'Scegli dal repertorio pubblico o inserisci il nome Github (utente o organizzazione) e il nome del repository di un repository ospitato su Github, qui sotto. ' +
-      'La tua repository forked sarà disponibile dal menu di Github.',
+      'La tua forked repository sarà disponibile dal menu di Github.',
   },
   forkRepositoryGithubText: { text: 'Fork Repository Github' },
   forkRepertoireSummary: { text: 'Repertorio pubblico' },
@@ -403,7 +407,7 @@ export const lang = {
   forkRepertoireOrganizationLabel: { text: 'Organizzazione:' },
   forkRepertoireOrganizationOption: { text: "Scegli un'organizzazione GitHub..." },
   forkRepertoireRepositoryLabel: { text: 'Repository:' },
-  forkRepertoireRepositoryOption: { text: 'Scegli una codice...' },
+  forkRepertoireRepositoryOption: { text: 'Scegli un codice...' },
   forkRepositoryInputName: { placeholder: 'Nome utente o organizzazione Github' },
   forkRepositoryInputRepoOption: { text: 'Scegli un repository' },
   forkRepositoryToSelectorText: { text: 'Fork a: ' },
@@ -418,14 +422,14 @@ export const lang = {
   },
 
   // CodeMirror editor
-  selectTagNameForEnclosure: { text: "Acchiudere con il nome dell'elemento" },
+  selectTagNameForEnclosure: { text: "Racchiudere con il nome dell'elemento" },
   selectTagNameForEnclosureOkButton: { value: 'OK' },
   selectTagNameForEnclosureCancelButton: { value: 'Annulla' },
 
   // restore Solid session overlay
   solidExplanation: {
     description:
-      'Solid è una piattaforma decentralizzata per dati collegati sociali. Accedi a Solid per creare annotazioni stand-off utilizzando dati collegati (RDF).',
+      'Solid è una piattaforma decentralizzata per memorizzare i dati personali degli utenti. Accedi a Solid per creare annotazioni stand-off utilizzando dati collegati (RDF).',
   },
   solidProvider: { description: 'Per favore, scegli un provider di identità Solid (IdP) o specifica il tuo.' },
   solidLoginBtn: { text: 'Accedi' },
@@ -440,9 +444,9 @@ export const lang = {
   },
 
   // annotation panel / annotation panel
-  annotationCloseButtonText: { text: 'Chiudi pannello di annotazione' },
-  hideAnnotationPanelButton: { description: 'Chiudi pannello di annotazione' },
-  closeAnnotationPanelButton: { description: 'Chiudi pannello di annotazione' },
+  annotationCloseButtonText: { text: 'Chiudi la finestra di annotazione' },
+  hideAnnotationPanelButton: { description: 'Nascondi la finestra di annotazione' },
+  closeAnnotationPanelButton: { description: 'Chiudi la finestra di annotazione' },
   annotationToolsButton: { text: 'Strumenti', description: 'Strumenti di annotazione' },
   annotationListButton: { text: 'Lista', description: 'Elenco delle annotazioni' },
   writeAnnotStandoffText: { text: 'Annotazione Web' },
@@ -469,14 +473,14 @@ export const lang = {
     descriptionToLocal: "Apri l'annotazione stand-off (RDF) in una nuova scheda",
   },
   makeInlineAnnotation: {
-    description: 'Clicca per annotazione in-linea',
+    description: 'Clicca per annotazione in-line',
     descriptionCopy: "Copia l'xml:id di <annot> negli appunti",
   },
   pageAbbreviation: { text: 'p.' },
   elementsPlural: { text: 'elementi' },
   askForLinkUrl: { text: "Inserisci l'URL a cui collegare" },
   drawLinkUrl: { text: 'Apri in una nuova scheda' },
-  askForDescription: { text: 'Inserisci una descrizione testuale da applicare' },
+  askForDescription: { text: 'Inserisci una descrizione' },
   maxNumberOfAnnotationAlert: {
     text1: 'Il numero di elementi <annot> supera il valore massimo configurabile di "Numero massimo di annotazioni"',
     text2:
@@ -557,8 +561,8 @@ export const lang = {
 
   // Drag select / Selezione tramite trascinamento
   dragSelection: {
-    text: 'Selezione tramite trascinamento',
-    description: 'Seleziona gli elementi nello spartito con il trascinamento del mouse',
+    text: 'Seleziona tramite trascinamento',
+    description: 'Seleziona gli elementi nello spartito trascinandole col mouse',
   },
   dragSelectNotes: {
     text: 'Seleziona le note',
@@ -629,7 +633,7 @@ export const lang = {
   },
   scrollFollowMidiPlayback: {
     text: 'Scorrimento automatico',
-    description: 'Scorri il pannello dello spartito durante la riproduzione MIDI sulla pagina corrente',
+    description: 'Scorri la finestra dello spartito durante la riproduzione MIDI sulla pagina corrente',
   },
   pageFollowMidiPlayback: {
     text: 'Cambio pagina automatico',
@@ -638,7 +642,7 @@ export const lang = {
   highlightCurrentlySoundingNotes: {
     text: 'Evidenzia note attualmente suonate',
     description:
-      'Evidenzia visivamente le note attualmente suonate nel pannello di notazione durante la riproduzione MIDI',
+      'Evidenzia visivamente le note attualmente suonate nella finestra di notazione durante la riproduzione MIDI',
   },
 
   selectMidiExpansion: {
@@ -754,8 +758,8 @@ export const lang = {
     description: 'Mostra le annotazioni nello spartito',
   },
   showAnnotationPanel: {
-    text: 'Mostra pannello annotazioni',
-    description: 'Mostra il pannello delle annotazioni',
+    text: 'Mostra la finestra annotazioni',
+    description: 'Mostra la finestra delle annotazioni',
   },
   annotationDisplayLimit: {
     text: 'Numero massimo di annotazioni',
@@ -764,16 +768,16 @@ export const lang = {
 
   // Facsimile / Facsimile
   titleFacsimilePanel: {
-    text: 'Pannello di facsimile',
+    text: 'Finestra di facsimile',
     description: "Mostra le immagini di facsimile dell'edizione originale, se disponibili",
   },
   showFacsimilePanel: {
-    text: 'Mostra il pannello di facsimile',
+    text: 'Mostra la finestra di facsimile',
     description: "Mostra le immagini dello spartito dell'edizione originale fornite nell'elemento di facsimile",
   },
   selectFacsimilePanelOrientation: {
-    text: 'Posizione del pannello di facsimile',
-    description: 'Seleziona la posizione del pannello di facsimile rispetto allo spartito',
+    text: 'Posizione della finestra di facsimile',
+    description: 'Seleziona la posizione della finestra di facsimile rispetto allo spartito',
     labels: ['sinistra', 'destra', 'alto', 'basso'],
   },
   facsimileZoomInput: {
@@ -841,7 +845,7 @@ export const lang = {
   },
   lineWrapping: {
     text: 'Rientro righe',
-    description: 'Indica se le righe vengono spezzate a fine pannello',
+    description: 'Indica se le righe vengono spezzate alla fine della finestra',
   },
   lineNumbers: {
     text: 'Numeri di riga',