+++ app/static/lang/lang.ja.js
@@ -159,6 +159,7 @@ export const lang = {
   staffDownText: { text: '要素を1段階下げる' },
   increaseDurText: { text: '音価を増やす' },
   decreaseDurText: { text: '音価を減らす' },
+  toggleDotsText: { text: 'トグル・ドッティング' },
   cleanAccidText: { text: '@accid.gesを確認' },
   renumberMeasuresTestText: { text: '小節を再番号付け（テスト）' },
   renumberMeasuresExecText: { text: '小節を再番号付け（実行）' },
@@ -167,7 +168,12 @@ export const lang = {
   reRenderMeiVerovio: { text: 'Verovioで再レンダリング' },
   addFacsimile: { text: 'ファクシミリを追加' },
   ingestFacsimileText: { text: 'ファクシミリ取り込み' },
+
+  // INSERT MENU
   insertMenuTitle: { text: '挿入' },
+  addNoteText: { text: '音符を追加' },
+  convertNoteToRestText: { text: '音符 <=> 休符' },
+  toggleChordText: { text: '音符 <=> 和音' },
   addDoubleSharpText: { html: 'ダブルシャープ &#119082;' },
   addSharpText: { html: 'シャープ &#9839;' },
   addNaturalText: { html: 'ナチュラル &#9838;' },