+++ app/static/lang/lang.uk.js
@@ -156,6 +156,7 @@ export const lang = {
   staffDownText: { text: 'Перемістити на 1 лінію вниз' },
   increaseDurText: { text: 'Збільшити тривалість' },
   decreaseDurText: { text: 'Зменшити тривалість' },
+  toggleDotsText: { text: 'перемикання пунктиру' },
   cleanAccidText: { text: 'Перевірити @accid.ges' },
   renumberMeasuresTestText: { text: 'Перенумерувати такти (тест)' },
   renumberMeasuresExecText: { text: 'Перенумерувати такти (виконати)' },
@@ -167,6 +168,9 @@ export const lang = {
 
   // INSERT MENU ITEM
   insertMenuTitle: { text: 'Вставити' },
+  addNoteText: { text: 'Додати ноту' },
+  convertNoteToRestText: { text: 'Нота(и) <=> пауза(и)' },
+  toggleChordText: { text: 'Нота(и) <=> акорд' },
   addDoubleSharpText: { html: 'Подвійний діез &#119082;' },
   addSharpText: { html: 'Дієз &#9839;' },
   addNaturalText: { html: 'Натураль &#9838;' },