+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # mei-friend-online CHANGELOG.md
+
+### 1.0.5 patch
+* Add Ukrainian and Esperanto language packs (still to be proof-read)
+* Keyboard shortcut (`CTRL + ALT + SPACE`) to switch focus between notation and encoding panel (#55)
+* Keyboard shortcut (`N`) to duplicate a selected chord, note or rest, or insert a new note into an empty layer (#57)
+* Keyboard shortcut (`SHIFT + N`) to convert selected notes to rests and rests to notes (#57)
+* Keyboard shortcut (`.`) to add or remove `dots="1"` to notes, chords, etc.
+* Keyboard shortcut (`C`) to embed selected notes inside a new chord element, and vice versa (remove chord and leave notes)
+
 ### 1.0.4 patch
 * Tweak CSS select and translations (German)
+* Fix typo at forking via URL parameters
 
 ### 1.0.3 patch
 * Enable navigation link to generated MAO objects (identify object annotations)
@@ -13,6 +23,7 @@
 * Show warning message when clicking on or drag-selecting elements without xml:ids (#10)
 * Drag'n'drop on CodeMirror editor now shows drag overlay screen
 * Take language from browser locale
+
 ### 1.0.1 patch
 * Reduce mei-friend-specific Verovio defaults
 * Add auto-translated draft of Japanese language pack